Reuse subjects cache inside ModelsTree
