+++ packages/itwin/tree-widget/package.json
@@ -61,10 +61,10 @@
     "@itwin/itwinui-icons-react": "^2.8.0",
     "@itwin/itwinui-illustrations-react": "^2.1.0",
     "@itwin/itwinui-react": "^3.5.0",
-    "@itwin/presentation-core-interop": "^0.1.2",
+    "@itwin/presentation-core-interop": "^0.2.0",
     "@itwin/presentation-hierarchies": "^0.1.1",
     "@itwin/presentation-hierarchies-react": "^0.2.0",
-    "@itwin/presentation-shared": "^0.1.1",
+    "@itwin/presentation-shared": "^0.2.0",
     "@itwin/unified-selection": "^0.4.1",
     "classnames": "^2.3.1",
     "i18next": "^10.2.2",
